Split API functions to their own files
